Use SDL_GetTicks64() in place of SDL_GetTicks()
As recommended by the SDL documentation:
"This function is not recommended as of SDL 2.0.18; use SDL_GetTicks64()
instead, where the value doesn't wrap every ~49 days."
https://wiki.libsdl.org/SDL2/SDL_GetTicks

+++ src/game/gfx.hpp
@@ -282,7 +282,7 @@ struct Gfx
 	bool running;
 	bool spectatorFullscreen, doubleRes;
 
-	Uint32 lastFrame;
+	uint64_t lastFrame;
 	unsigned menuCycles;
 	int windowW, windowH;
 	int prevMag; // Previous magnification used for drawing
